+++ src/widgets/header/constants/navItem.ts
@@ -1,0 +1,7 @@
+export const NAV_ITEMS = [
+  { name: 'About', href: '/#about' },
+  { name: 'Project', href: '/#project' },
+  { name: 'Activity', href: '/#activity' },
+  { name: 'FAQ', href: '/#faq' },
+  { name: 'Join Us', href: '/joinus' }
+];
